fix up the unit tests
