+++ library/src/androidTest/java/org/xmtp/android/library/ClientTest.kt
@@ -1,9 +1,15 @@
 package org.xmtp.android.library
 
+import androidx.test.ext.junit.runners.AndroidJUnit4
 import org.junit.Assert.assertEquals
 import org.junit.Test
+import org.junit.runner.RunWith
 import org.xmtp.android.library.messages.PrivateKeyBuilder
+import org.xmtp.android.library.messages.PrivateKeyBundleV1Builder
+import org.xmtp.android.library.messages.generate
+import org.xmtp.proto.message.contents.PrivateKeyOuterClass
 
+@RunWith(AndroidJUnit4::class)
 class ClientTest {
     @Test
     fun testTakesAWallet() {
@@ -18,6 +24,20 @@ class ClientTest {
         assertEquals(1, client.privateKeyBundleV1?.preKeysList?.size)
         val preKey = client.privateKeyBundleV1?.preKeysList?.get(0)
         assert(preKey?.publicKey?.hasSignature() ?: false)
+    }
+
+    @Test
+    fun testSerialization() {
+        val wallet = PrivateKeyBuilder()
+        val v1 =
+            PrivateKeyOuterClass.PrivateKeyBundleV1.newBuilder().build().generate(wallet = wallet)
+        val encodedData = PrivateKeyBundleV1Builder.encodeData(v1)
+        val v1Copy = PrivateKeyBundleV1Builder.fromEncodedData(encodedData)
+        val client = Client().buildFrom(v1Copy)
+        assertEquals(
+            wallet.address,
+            client.address
+        )
     }
 
     @Test
